Use an LLM to "reword" all the bad CSV rows (#168)
This script / .NET console app:

* Opens `classified.csv`
* For each row, ask an LLM how to reword so it's not bad
* Save the rows in a new `output.csv`

I then manually added the rows to `classified.csv`, sorting alphabetically.

I also used SemanticKernel, which might be overkill for this. But it could easily be swapped out to use different LLMs.

TODO:

* The code is non-optimal, it could allocate less
* It's just synchronous

I was honestly OK for this code to be slow, as I was scared what it would do to my OpenAI balance.

The rows added here cost $0.04 with `gpt-3.5-turbo`. They honestly seem good enough?